# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
        "/api/peliculas": {
            "get": {
                "tags": [
                    "Pelicula"
                ],
                "summary": "Mostar toda las peliculas",
                "operationId": "App\\Http\\Controllers\\PeliculaController::index",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar toda las Peliculas ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Pelicula"
                ],
                "summary": "Registrar",
                "operationId": "App\\Http\\Controllers\\PeliculaController::store",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "a3fb6",
                                    "descripcion": "Jessica Smith",
                                    "year": 2020,
                                    "categoria_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registrar Pelicula ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/peliculas/{id}": {
            "get": {
                "tags": [
                    "Pelicula"
                ],
                "summary": "Show",
                "operationId": "App\\Http\\Controllers\\PeliculaController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buscar una pelicula ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Pelicula"
                ],
                "summary": "Editar una pelicula",
                "operationId": "App\\Http\\Controllers\\PeliculaController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcion",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 0,
                                    "nombre": "a3fb6",
                                    "descripcion": "Jessica Smith",
                                    "year": 2020,
                                    "categoria_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Editar una pelicula ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pelicula"
                ],
                "summary": "Eliminar una pelicula",
                "operationId": "App\\Http\\Controllers\\PeliculaController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar una pelicula ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}